Install LLVM + clang prior to generating Ada bindings
Some LLVM headers are now generated at install time, so we can't
use the header in the sources as is to get Ada bindings for all of
them. Build and install LLVM and clang prior to generating the bindings,
and use the installed headers instead of the ones in the sources.

TN: V418-010
Change-Id: I5735d25b8ac418a168d8e81719a456c514e08b30