Properly compute size to move in unconstrained case
This is another occurrence of one of the issues that came up with
W413-008.  If there's a variant record that's used as a component of
an outer object, that component's size will be the maximum for that
variant record, but we need to be sure that we only copy the actual
size of the object to avoid reading or writing undefined bytes.

Fix for eng/toolchain/gnat-llvm/-/issues/1.