Add support for -fsanitize=fuzzer,address
The -fsanitize option enables various sanitizers in Clang. We need
address sanitizer and libfuzzer in GNATfuzz, so for now we only support
those two options in GNAT-LLVM. Their effect is twofold:

1. Add an instrumentation pass to the optimization pipeline.
2. Tag functions with additional attributes used during optimization.

Note that we only handle the compilation aspects of the sanitizers; we
don't need to do anything at link time because we use Clang to drive the
linker, which already understands the relevant options.

Fixes eng/toolchain/gnat-llvm#26.