Fix handling of atomic access to addresses
Some of GCC's atomic builtins come in two flavors: one takes certain
parameters as values, and the other takes pointers. We currently handle
both at the same time by checking parameter types and dereferencing if
they're addresses or access types. However, this approach of "maybe
dereferencing" parameters doesn't work when the target of the atomic
operation is an address: in this case, it's unclear whether address
parameters are values or pointers.

This commit splits Emit_And_Maybe_Deref into two distinct functions,
Emit_And_Deref and Emit_And_Convert, and chooses one or the other based
on the atomic operation's name (which determines whether or not it's one
of the generic variants taking pointer operands).

Issue: eng/toolchain/gnat-llvm#64