Stop using --target=llvm to build the runtime
In addition to preparing us for the removal of the "llvm" target, this
change allows us to drop the explicit configuration of the "--sysroot"
argument when targeting cross Linux; gprbuild now handles it for us. The
same applies to other cross targets that we may want to add.

Fixes eng/toolchain/gnat-llvm#149.