Use Clang to compile C and C++ code during bootstrap
We used to enforce g++ as the C++ compiler because of issues with some
Linux distributions. That part of the project configuration isn't
applied anymore now that we've stopped using "--target=llvm"; instead of
restoring it, this commit switches the bootstrap build to Clang.

The issue that we had with Clang was likely caused by it falling back to
the system C++ compiler's runtime. We avoid the problem by explicitly
pointing Clang to our own GCC, making it use our GCC's C++ runtime
instead of the system's.

Issue: eng/toolchain/gnat-llvm#149