Make the stack executable for trampolines on Linux
Trampolines require the stack to be executable. When we still used the
GNU linker, it would automatically leave the stack executable if any of
the object files didn't contain a section called .note.GNU-stack (which
is the case when LLVM emits trampolines). The LLVM linker that we've
recently switched to always makes the stack non-executable for security
reasons, so we need to call __enable_execute_stack on Linux explicitly
when using trampolines.

Issues: eng/toolchain/llvm-project#14, eng/toolchain/gnat-llvm#133