Rewrite check for DISubrangeType
My earlier compile-time approach to checking for DISubrangeType did
not in fact work -- I tested it, but I must have made some mistake,
because when testing a subsequent patch I found out that the
compilation failed with an unpatched LLVM.

This patch fixes the problem by adding a checking script.  The script
is written to be extensible as more checks like this will be needed.