Special case "main" in DWARF
GCC has a special case for the "main" function, omitting the logical
name and just emitting "main" as the name.  This patch replicates this
behavior for gnat-llvm.  This enables a few gdb tests to run a bit
further.

While we could work around this difference in gdb, it seemed simpler
to implement it here.