Handle self-referential types
I noticed that a self-referential record type like this:

   type T1 is record
      Value : Natural;
      Next : access all T1;
   end record;

would end up with the type of 'Next' being a pointer-to-unspecified
type.  This is a consequence of this code in Create_Type_Data:

      elsif Is_Being_Elaborated (TE) then
         return DI_Create_Unspecified_Type (Name);

This patch cleans this up by creating the LLVM debuginfo type earlier
and then later updating its members.  The debug metadata cache is used
to ensure that recursive calls pick up this type.