Remove elaboration check from Create_Type_Data
The tasking code in gdb was not working, and I tracked this down to
this DWARF:

 <1><9e49>: Abbrev Number: 19 (DW_TAG_unspecified_type)
    <9e4a>   DW_AT_name        : (indirect string, offset: 0x26cb): system__tasking__task_id

This particular CU has two copies of this type: one as an access type,
and the one shown -- an unspecified type.  This is being emitted due
to the elaboration check in Create_Type_Data, since this type is
recursive.

I think this elaboration check was an earlier way to try to cope with
recursive types.  I tried to solve this earlier, see commit 4a857a39
("Handle self-referential types"), but there I didn't consider the
case where the access type is seen before the record type.

Removing this check entirely fixes the problem and allows gdb's
tasking features to work.

However, doing this naively results in another bug -- in Ada, a
pointer type can be recursive.  While this isn't useful, the compiler
should not crash in this case.  So this patch also adds a check for
recursive access types and turns these into pointer-to-unspecified.

eng/toolchain/gnat-llvm#153